requirements.txt
asgiref==3.8.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
confection==0.1.5
cymem==2.0.8
distlib==0.3.9
Django==5.1.3
djangorestframework==3.15.2
en-core-med7-lg @ https://huggingface.co/kormilitzin/en_core_med7_lg/resolve/main/en_core_med7_lg-any-py3-none-any.whl#sha256=2c8ce371b7ee8b3c42cce7c05031ab165e6e88cc3ed136797eb3c492af96001f
filelock==3.16.1
idna==3.10
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.10
numpy==1.21.6
packaging==24.2
pathlib_abc==0.1.1
pathy==0.11.0
pipenv==2024.4.0
platformdirs==4.3.6
preshed==3.0.9
pydantic==1.10.19
requests==2.32.3
smart-open==6.4.0
spacy==3.4.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.4.8
thinc==8.1.12
tqdm==4.67.0
typer==0.4.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
wasabi==0.10.1
